+++ objectified-ui/public/WHATS_NEW.md
@@ -28,6 +28,7 @@ We continue to improve the platform based on your feedback with improvements and
     - Adds OAuth2 security options
     - Adds OpenID security options
     - Adds TLS security options
+    - Adds custom security schema options
   - Dragging a property to a variable chip in the header will apply that property's primitive settings to the bound variable
 - OpenAPI Support:
   - Now at 100% support for class and properties definitions